pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ghostcontext"
version = "0.1.0"
description = "Local-first OpenAI-compatible proxy with ChromaDB memory and file logs"
readme = "README.md"
requires-python = ">=3.11"
license = { file = "LICENSE" }
authors = [{ name = "GhostContext contributors" }]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "chromadb>=0.5.0",
  "openai>=1.55.0",
  "python-dotenv>=1.0.1",
  "pydantic-settings>=2.6.0",
]

[project.optional-dependencies]
dev = ["httpx>=0.27.0", "ruff>=0.8.0"]

[project.scripts]
ghostcontext = "ghostcontext.main:run"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"
